# specopenapi.yaml
openapi: 3.0.1
info:
  title: Delilah Restó
  description: Documentación de la API REST de Delilah Restó.
  contact:
    email: delilahdev@gmail.com
  version: 1.0.0
...
servers:
- url: https://delilah.resto.com.ar/
- url: http://delilah.resto.com.ar/
...
paths:
...
  /pedidos:
    post:
      tags:
      - pedido
      summary: Subir un pedido nuevo
      description: Se sube el pedido de un usuario con el/los plato/s requeridos y la información de la transacción.
      operationId: subirPedido
      requestBody:
        description: Objeto con la información del pedido de un determinado cliente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pedido'
        required: true
      responses:
        200:
          description: Operacion exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExitoPedido'
        401:
          description: No se encontró ningún token de autorización
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        422:
          description: Formulario del pedido incompleto y/o con errores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exito'
      security:
      - bearerAuth: []
      x-codegen-request-body-name: body
...
  /usuarios/registro:
    post:
      tags:
      - usuario
      summary: Crear un usuario nuevo
      description: Para poder registrarse, el usuario debe ingresar username, nombre completo, email, telefono, direccion y password.
      operationId: crearUsuario
      requestBody:
        description: Objeto nuevo usuario
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
        required: true
      responses:
        200:
          description: Registro exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Exito'
        422:
          description: Formulario de registro incompleto y/o con errores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:            
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Usuario:
      type: object
      properties:
        username:
          type: string
        nombre_completo:
          type: string
        email:
          type: string
        telefono:
          type: string
        direccion:
          type: string
        password:
          type: string
          format: password
    Cantidad:
      type: object
      properties:
        id_plato:
          type: integer
        cantidad:
          type: integer
    Pedido:
      type: object
      properties:
        id_usuario:
          type: integer
        descripcion:
          type: array
          items:
            $ref: '#/components/schemas/Cantidad'
        tipo_pago:
          type: string
          description: Tipos de pago disponbles
          enum:
          - efectivo
          - debito
          - credito
        total:
          type: number
...
    Error:
      type: object
      properties:
        error:
          type: string
    Exito:
      type: object
      properties:
        mensaje:
          type: string
    ExitoPedido:
      type: object
      properties:
        mensaje:
          type: string
        id_pedido:
          type: integer
...